Move errors to engine module
