+++ radix-engine/src/engine/call_frame.rs
@@ -16,10 +16,7 @@ use scrypto::values::*;
 
 use self::LazyMapState::{Committed, Uncommitted};
 use self::LoadedSNodeState::{Borrowed, Consumed, Static};
-use crate::engine::track::{SubstateValue, TrackError};
-use crate::engine::ComponentObjectRefs;
 use crate::engine::*;
-use crate::errors::*;
 use crate::ledger::*;
 use crate::model::*;
 use crate::wasm::*;
