+++ radix-engine/src/engine/component_objects.rs
@@ -2,7 +2,7 @@ use scrypto::engine::types::*;
 use scrypto::rust::collections::*;
 use scrypto::rust::vec::Vec;
 
-use crate::errors::*;
+use crate::engine::*;
 use crate::model::*;
 
 #[derive(Debug)]
