+++ radix-engine/src/engine/mod.rs
@@ -1,5 +1,6 @@
 mod call_frame;
 mod component_objects;
+mod errors;
 mod id_allocator;
 mod id_validator;
 mod receipt;
@@ -10,10 +11,13 @@ mod track;
 
 pub use call_frame::{CallFrame, SNodeState};
 pub use component_objects::*;
+pub use errors::*;
 pub use id_allocator::*;
 pub use id_validator::*;
 pub use receipt::*;
 pub use runtime::RadixEngineScryptoRuntime;
 pub use substate_receipt::{CommitReceipt, SubstateOperation, SubstateOperationsReceipt};
 pub use system_api::SystemApi;
-pub use track::{Address, BorrowedSNodes, SubstateParentId, SubstateUpdate, SubstateValue, Track};
+pub use track::{
+    Address, BorrowedSNodes, SubstateParentId, SubstateUpdate, SubstateValue, Track, TrackError,
+};
