+++ radix-engine/src/engine/receipt.rs
@@ -9,7 +9,7 @@ use scrypto::rust::vec::Vec;
 use scrypto::values::*;
 
 use crate::engine::CommitReceipt;
-use crate::errors::RuntimeError;
+use crate::engine::RuntimeError;
 use crate::model::*;
 
 /// Represents a transaction receipt.
