+++ radix-engine/src/engine/system_api.rs
@@ -4,7 +4,7 @@ use scrypto::rust::string::String;
 use scrypto::rust::vec::Vec;
 use scrypto::values::*;
 
-use crate::errors::*;
+use crate::engine::*;
 use crate::model::*;
 
 pub trait SystemApi {
