+++ radix-engine/src/lib.rs
@@ -8,13 +8,11 @@ compile_error!("Feature `std` and `alloc` can't be enabled at the same time.");
 
 /// Radix Engine implementation.
 pub mod engine;
-/// Radix Engine errors.
-pub mod errors;
-/// Radix ledger abstraction.
+/// Radix Engine state abstraction.
 pub mod ledger;
-/// Radix Engine transaction and state models.
+/// Radix Engine models.
 pub mod model;
-/// Transaction builder, validator and executor.
+/// Radix Engine transaction abstraction.
 pub mod transaction;
 /// Wasm validation, instrumentation and execution.
 pub mod wasm;
