+++ radix-engine/src/model/auth_converter.rs
@@ -1,8 +1,3 @@
-use crate::model::method_authorization::{
-    HardAuthRule, HardCount, HardDecimal, HardProofRule, HardProofRuleResourceList,
-    HardResourceOrNonFungible,
-};
-use crate::model::MethodAuthorization;
 use sbor::any::Value;
 use sbor::*;
 use scrypto::engine::types::*;
@@ -13,6 +8,12 @@ use scrypto::resource::{
 };
 use scrypto::rust::vec::Vec;
 use scrypto::types::ScryptoType;
+
+use crate::model::method_authorization::{
+    HardAuthRule, HardCount, HardDecimal, HardProofRule, HardProofRuleResourceList,
+    HardResourceOrNonFungible,
+};
+use crate::model::MethodAuthorization;
 
 fn soft_to_hard_decimal(schema: &Type, soft_decimal: &SoftDecimal, dom: &Value) -> HardDecimal {
     match soft_decimal {
