+++ radix-engine/src/model/system.rs
@@ -1,8 +1,9 @@
-use crate::engine::SystemApi;
 use sbor::DecodeError;
 use scrypto::buffer::scrypto_decode;
 use scrypto::core::SystemFunction;
 use scrypto::values::ScryptoValue;
+
+use crate::engine::SystemApi;
 
 #[derive(Debug, Clone, PartialEq)]
 pub enum SystemError {
