+++ radix-engine/src/model/transaction_processor.rs
@@ -1,8 +1,3 @@
-use crate::engine::{IdAllocator, IdSpace, SystemApi};
-use crate::errors::RuntimeError;
-use crate::errors::RuntimeError::ProofNotFound;
-use crate::model::worktop::WorktopMethod;
-use crate::model::{ValidatedInstruction, ValidatedTransaction};
 use scrypto::call_data;
 use scrypto::component::PackageFunction;
 use scrypto::core::SNodeRef;
@@ -13,6 +8,10 @@ use scrypto::rust::collections::HashMap;
 use scrypto::rust::string::ToString;
 use scrypto::rust::vec::Vec;
 use scrypto::values::*;
+
+use crate::engine::{IdAllocator, IdSpace, RuntimeError, RuntimeError::ProofNotFound, SystemApi};
+use crate::model::worktop::WorktopMethod;
+use crate::model::{ValidatedInstruction, ValidatedTransaction};
 
 pub struct TransactionProcessor {
     transaction: ValidatedTransaction,
