+++ radix-engine/src/transaction/error.rs
@@ -2,7 +2,7 @@ use sbor::describe::Type;
 use scrypto::engine::types::*;
 use scrypto::rust::string::String;
 
-use crate::errors::RuntimeError;
+use crate::engine::RuntimeError;
 
 /// Represents an error when parsing arguments.
 #[derive(Debug, Clone)]
