+++ radix-engine/tests/bucket.rs
@@ -2,7 +2,7 @@
 pub mod test_runner;
 
 use crate::test_runner::TestRunner;
-use radix_engine::errors::*;
+use radix_engine::engine::*;
 use radix_engine::ledger::*;
 use radix_engine::model::{BucketError, ResourceContainerError};
 use radix_engine::transaction::*;
