+++ simulator/src/resim/error.rs
@@ -1,6 +1,6 @@
 use std::io;
 
-use radix_engine::errors::*;
+use radix_engine::engine::*;
 use radix_engine::transaction::*;
 use radix_engine::wasm::*;
 use sbor::*;
